Task 2/requirements.txt
kafka-python==2.0.2
scikit-learn==1.2.2
joblib==1.2.0
pandas==2.1.2
numpy==1.26.4
pyspark==3.4.1